Add initial structtags package.
Create a helper function for mapping a tftypes.Object to a struct using
the struct tags to map to tftypes.Values within the object.

This is not perfect. We could have smarter types support (though maybe
that belongs in As? :/) and needs some smarter struct tags to e.g.
specify what happens in the event of unknown values so users can specify
that unknown values get treated as null or zero values rather than
errors, if they want to ignore that complexity. It also needs the
ability for lists, tuples, and sets to be converted into something other
than []tftypes.Value, which users of this helper probably don't want,
and ditto for maps and objects into something other than
map[string]tftypes.Value, which seems similarly like a thing users of
this helper would not want. More robust testing is also in order for the
variety of edge cases. (I've neatly danced around nil for the moment, for
example.)

But the test passes, and that's good enough for tonight.