Add asgotypes package. (#1)
* Add asgotypes package.

Add a package that will allow unpacking a `tftypes.Value` without
knowing anything about its type.

This is an alternative to what was proposed in
hashicorp/terraform-plugin-go#50.

* Rename from Default to GoPrimitive.

As @paultyng pointed out, the naming of Default made it sound like we
were working with Terraform defaults, not a default type mapping. Rename
the package and the main type to make it clearer we're mapping to Go
primitives.